Refactored Dexie.d.ts (prepare for refactoring classes) Dexie.Promise is experimentally removed in favour of the standard Promise. Reason is to cope with Typescript standard library changes without having to catch-up with them. However, this could break existing codes that uses Promise.finally(), Promise.catch(errName, err) etc. So we might need to put it back later on for backward compatibility. Let's see.
