+++ src/interfaces/dexie-event.d.ts
@@ -1,0 +1,6 @@
+export interface DexieEvent {
+  subscribers: Function[];
+  fire(...args:any[]): any;
+  subscribe(fn: (...args:any[]) => any): void;
+  unsubscribe(fn: (...args:any[]) => any): void;
+}
