+++ src/interfaces/index-spec.d.ts
@@ -1,0 +1,9 @@
+export interface IndexSpec {
+  name: string;
+  keyPath: string | Array<string>;
+  unique: boolean;
+  multi: boolean;
+  auto: boolean;
+  compound: boolean;
+  src: string;
+}
