+++ src/interfaces/table-schema.d.ts
@@ -1,0 +1,8 @@
+import { IndexSpec } from "./index-spec";
+
+export interface TableSchema {
+  name: string;
+  primKey: IndexSpec;
+  indexes: IndexSpec[];
+  mappedClass: Function;
+}
