+++ src/interfaces/transaction.d.ts
@@ -1,0 +1,13 @@
+export interface Transaction {
+  active: boolean;
+  db: Dexie;
+  mode: string;
+  idbtrans: IDBTransaction;
+  tables: { [type: string]: Table<any, any> };
+  storeNames: Array<string>;
+  on: TransactionEvents;
+  abort(): void;
+  table(tableName: string): Table<any, any>;
+  table<T>(tableName: string): Table<T, any>;
+  table<T, Key>(tableName: string): Table<T, Key>;
+}
