+++ src/types/then-shortcut.d.ts
@@ -1,0 +1,1 @@
+export type ThenShortcut<T,TResult> =  (value: T) => TResult | PromiseLike<TResult>;
