+++ src/types/transaction-mode.d.ts
@@ -1,0 +1,1 @@
+export type TransactionMode = 'r' | 'r!' | 'r?' | 'rw' | 'rw!' | 'rw?';
